feat(agents): register pi as a selectable ACP agent
getAgent() returns undefined for any name outside KNOWN_AGENT_NAMES, so a
config profile naming "pi" would have resolved to no adapter. Add pi to the
known names and give it a real ACP registry entry rather than letting it fall
through to DEFAULT_ENTRY.

The per-agent registry table moves to its own module: adapter.ts was a
grandfathered oversized file at 629 lines and the new row pushed it past the
ratchet. Extracting the table drops adapter.ts to 590 lines, back under the
600-line limit, so the oversized baseline falls from 11 files to 10.

pi is reached through the third-party pi-acp bridge that acpx spawns via npx.
It advertises model selection as an ACP config option, so all three tiers are
selectable; the context ceiling is set conservatively to match pi's own default
for models whose metadata it cannot resolve.
